+++ CHANGELOG.md
@@ -7,7 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.1.1](https://github.com/bfg-s/dto/compare/1.1.0...1.1.1) - 2024-12-09
 * Fixed `make:dto-docs` command, now it generates documentation and for `lazy` properties and for computed properties
 * Fixed reflexion `fromArray...` and `toArray...` now is not computed properties
-
+* Method for getting http client
+* Property for getting by post request in anything constructor
 
 ## [1.1.0](https://github.com/bfg-s/dto/compare/1.0.0...1.1.0) - 2024-12-08
 * `toArray` revert back properties with `DtoName` attribute to their attribute name